GOB: move draw code into seperate folder for draw
move all draw related source code files into an folder for better code overview, like for save or sound
