+++ engines/gob/demos/demoplayer.cpp
@@ -33,7 +33,7 @@
 #include "gob/demos/demoplayer.h"
 #include "gob/global.h"
 #include "gob/util.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/inter/inter.h"
 #include "gob/video/videoplayer.h"
 #include "gob/sound/sound.h"
