+++ engines/gob/draw/draw.cpp
@@ -29,7 +29,7 @@
 #include "common/str.h"
 
 #include "gob/gob.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/dataio.h"
