+++ engines/gob/draw/draw_fascin.cpp
@@ -25,7 +25,7 @@
  *
  */
 
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/global.h"
 #include "gob/inter/inter.h"
