+++ engines/gob/draw/draw_v1.cpp
@@ -30,7 +30,7 @@
 #include "graphics/cursorman.h"
 
 #include "gob/gob.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/game.h"
