+++ engines/gob/draw/draw_v7.cpp
@@ -31,7 +31,7 @@
 #include "graphics/wincursor.h"
 
 #include "gob/dataio.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/global.h"
 #include "gob/hotspots.h"
