+++ engines/gob/hotspots.cpp
@@ -29,7 +29,7 @@
 
 #include "gob/hotspots.h"
 #include "gob/global.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/script.h"
 #include "gob/inter/inter.h"
