+++ engines/gob/init_v1.cpp
@@ -30,7 +30,7 @@
 #include "gob/gob.h"
 #include "gob/init.h"
 #include "gob/global.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 
 namespace Gob {
