+++ engines/gob/init_v2.cpp
@@ -30,7 +30,7 @@
 #include "gob/gob.h"
 #include "gob/init.h"
 #include "gob/global.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/video/video.h"
 
 namespace Gob {
