+++ engines/gob/inter/inter.cpp
@@ -31,7 +31,7 @@
 #include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/util.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/expression.h"
 #include "gob/script.h"
