+++ engines/gob/inter/inter_geisha.cpp
@@ -37,7 +37,7 @@
 #include "gob/script.h"
 #include "gob/resources.h"
 #include "gob/game.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/video/video.h"
 #include "gob/cheater.h"
 #include "gob/save/saveload.h"
