+++ engines/gob/inter/inter_littlered.cpp
@@ -29,7 +29,7 @@
 #include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/util.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/script.h"
 #include "gob/hotspots.h"
