+++ engines/gob/inter/inter_v3.cpp
@@ -32,7 +32,7 @@
 #include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/dataio.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/script.h"
 #include "gob/resources.h"
