+++ engines/gob/inter/inter_v4.cpp
@@ -30,7 +30,7 @@
 #include "gob/gob.h"
 #include "gob/inter/inter.h"
 #include "gob/global.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/script.h"
 #include "gob/video/videoplayer.h"
