+++ engines/gob/minigames/geisha/diving.cpp
@@ -29,7 +29,7 @@
 
 #include "gob/global.h"
 #include "gob/palanim.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/video/video.h"
 #include "gob/decfile.h"
 #include "gob/anifile.h"
