+++ engines/gob/minigames/geisha/penetration.cpp
@@ -30,7 +30,7 @@
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/palanim.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/video/video.h"
 #include "gob/decfile.h"
 #include "gob/cmpfile.h"
