+++ engines/gob/mult.cpp
@@ -31,7 +31,7 @@
 #include "gob/mult.h"
 #include "gob/global.h"
 #include "gob/util.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/script.h"
 #include "gob/palanim.h"
