+++ engines/gob/mult_v1.cpp
@@ -32,7 +32,7 @@
 #include "gob/mult.h"
 #include "gob/global.h"
 #include "gob/util.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/game.h"
 #include "gob/script.h"
 #include "gob/resources.h"
