+++ engines/gob/pregob/onceupon/parents.cpp
@@ -29,7 +29,7 @@
 #include "gob/global.h"
 #include "gob/dataio.h"
 #include "gob/palanim.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/video/video.h"
 
 #include "gob/sound/sound.h"
