+++ engines/gob/pregob/pregob.cpp
@@ -33,7 +33,7 @@
 #include "gob/surface.h"
 #include "gob/dataio.h"
 #include "gob/palanim.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/video/video.h"
 #include "gob/aniobject.h"
