+++ engines/gob/video/video_v6.cpp
@@ -33,7 +33,7 @@
 #include "gob/gob.h"
 #include "gob/video/video.h"
 #include "gob/util.h"
-#include "gob/draw.h"
+#include "gob/draw/draw.h"
 #include "gob/global.h"
 
 namespace Gob {
